requirements.txt
ffmpeg-python
whisper
tqdm
openai